+++ src/gui.js
@@ -91,7 +91,7 @@ modules.gui = '2025-March-21';
 
 // Declarations
 
-var SnapVersion = '10.6.2-250321-dev';
+var SnapVersion = '10.6.2';
 
 var IDE_Morph;
 var ProjectDialogMorph;