+++ HISTORY.md
@@ -4,7 +4,8 @@
 * **Notable Fixes:**
     * Thread halos: "afterglow" - display halos around running scripts for a couple of frames after their process has terminated (except for generic / custom event hats)
 
-### 2025-03-20
+### 2025-03-21
+* new dev version
 * blocks, objects, threads: "afterglow" - display halos around running scripts for a couple of frames after their process has terminated (except for generic / custom event hats)
 
 ## 10.6.1: