+++ HISTORY.md
@@ -2,11 +2,13 @@
 
 ## in development:
 * **Notable Fixes:**
-    * Thread halos: "afterglow" - display halos around running scripts for a couple of frames after their process has terminated (except for generic / custom event hats)
+    * Thread halos: "afterglow" - (again) display halos around running scripts for a couple of frames after their process has terminated (except for generic / custom event hats)
+    * fixed STRETCH COSTUME reporter to again also accept a 4-column list of pixel data 
 
 ### 2025-03-21
 * new dev version
 * blocks, objects, threads: "afterglow" - display halos around running scripts for a couple of frames after their process has terminated (except for generic / custom event hats)
+* threads: fixed STRETCH COSTUME reporter to again also accept a 4-column list of pixel data 
 
 ## 10.6.1:
 * **Notable Changes:**